src/requirements.txt
google-adk[a2a]
litellm
openai
